requirements.txt
fastapi==0.104.1
pydantic==2.5.0
pydantic-settings==2.1.0
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2
python-dotenv==1.0.0
uvicorn[standard]==0.24.0.post1